Refactor : presentation - domain 계층 분리
