+++ auth/src/main/java/com/eleven/logistics/auth/domain/entity/User.java
@@ -1,7 +1,6 @@
 package com.eleven.logistics.auth.domain.entity;
 
 import com.eleven.logistics.auth.domain.vo.Role;
-import com.eleven.logistics.auth.presentation.rest.dto.SignInRequestDto;
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -44,7 +43,7 @@ public class User {
                 matchesUsername(username)) {
             return this;
         }
-        throw new IllegalArgumentException("유효하지 않은 username 혹은 passoword");
+        throw new IllegalArgumentException("유효하지 않은 username 혹은 password");
     }
 
     private boolean matchesPassword(String password, PasswordEncoder passwordEncoder) {
