+++ auth/src/main/java/com/eleven/logistics/auth/infrastructure/repository/UserRepositoryImpl.java
@@ -2,11 +2,9 @@ package com.eleven.logistics.auth.infrastructure.repository;
 
 import com.eleven.logistics.auth.domain.entity.User;
 import com.eleven.logistics.auth.domain.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
